fix(embeds): degrade an unresolvable embed to a link instead of leaking its tag
An enabled provider that cannot resolve its input returned `None`, and the
authored tag was written back into the document verbatim. `<zenn url="…">`
reaches the browser as an unknown element: a self-closing one renders nothing
at all, so a link the author wrote disappears from the page with no diagnostic
anywhere. Nothing downstream catches it — the component resolver never sees
these tags.

An enabled provider now falls back to a plain link built from the URL it was
given, which needs no network and is identical in an offline build. A tag whose
URL is not safe to link to keeps its markup, because there is nothing better to
say, and a disabled provider is still left untouched.

The fallback carries no provider identity — no `--speakerdeck` modifier, no
data attribute. A renderer reports only "not mine", never why, so the tag that
reaches the fallback may be a real provider URL of an unrecognised shape or a
look-alike host such as `speakerdeck.com.evil.com`. Naming the provider would
let the second borrow that provider's styling, and `None` cannot tell them
apart.

Refs #1069.
