+++ docs/content/ja/built-in/embeds.md
@@ -147,6 +147,18 @@ GitHub issue、pull request、commit、discussion、gist の URL も、同じ
 
 favicon は対象ページ自身の `<link rel="icon">` から取り、無ければその origin の `/favicon.ico` へフォールバックします。サードパーティの favicon サービスへは接続しないので、カードを描画してもドキュメントページのリンク先が外部ホストへ伝わりません。
 
+## 埋め込みを解決できない場合
+
+プロバイダは認識できる入力に対してのみカードを描画します。有効なプロバイダがタグを解決できない場合（対象外のホスト、知らないパス形状など）、タグは未知の要素としてページに残るのではなく、素のリンクへ降格します。
+
+```html
+<a class="ox-embed-fallback" href="https://qiita.com/ubugeeei" target="_blank" rel="noopener noreferrer">https://qiita.com/ubugeeei</a>
+```
+
+リンク文言はタグの本文、次に `title`、最後に URL の順で決まります。フォールバックの class にプロバイダ名は含めません。レンダラは「自分のものではない」ことだけを伝え、ホストが違うのかパスだけが違うのかを区別しないため、名前を付けると偽装ホストがそのプロバイダのスタイルを借りられてしまうからです。
+
+リンクとして安全でない URL（HTTPS 以外のスキーム、埋め込み資格情報など）を持つタグは元のマークアップのまま残り、無効なプロバイダには一切手を触れません。
+
 ## パッケージマネージャタブ
 
 `embeds.pm` は 1 つの npm 風コマンドを、npm、pnpm、yarn、bun、vp（Vite+）向けのアクセシブルなタブグループへ展開します。
